+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* Added `WithConcurrentResultSets` option for `db.Query().Query()`
 * Added `DefaultValue` field to `table/options.Column` struct
 
 ## v3.124.1
@@ -44,7 +45,6 @@
 ## v3.118.1
 * Fixed connection timeout issue in topics writer
 * Supported `sql.Null*` from `database/sql` as query params in `toValue` func
-* Added `WithConcurrentResultSets` option for `db.Query().Query()`
 
 ## v3.118.0
 * Added support for nullable `Date32`, `Datetime64`, `Timestamp64`, and `Interval64` types in the `optional` parameter builder